emishows/pyproject.toml
[tool.poetry]
name = "emishows"
version = "0.1.0"
description = "emission shows 🎭"
readme = "README.md"
authors = ["radio-aktywne <it@radioaktywne.pl>"]
classifiers = ["License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"]
repository = "https://github.com/radio-aktywne/emishows"
documentation = "https://radio-aktywne.github.io/emishows"

[tool.poetry.dependencies]
python = "^3.9"
typer = { version = "^0.4", extras = ["all"] }
click = "~8.0.0"
djangorestframework = "^3.13"
Django = "~4.0.0"
psycopg2-binary = "^2.9"
django-cockroachdb = "~4.0.0"
django-filter = "^21.1"
caldav = "^0.8"
icalendar = "^4.0"
recurring-ical-events = { version = "^1.0", allow-prereleases = true }
tzdata = "*"
uvicorn = { version = "^0.17", extras = ["standard"] }
whitenoise = "^6.0"
pydantic = "^1.9"
httpx = "^0.22"

# dev

# test
pytest = { version = "^7.0", optional = true }

[tool.poetry.extras]
# need to do it that way until poetry supports dependency groups: https://github.com/python-poetry/poetry/issues/1644
dev = ["pytest"]
test = ["pytest"]
[tool.poetry.scripts]
# cli entry point
emishows = "emishows.__main__:cli"
emishows-manage = "emishows.manage:main"

[build-system]
# this should be there, see https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
